feat(refactor): conversion of all compedium data for JSON files and DB files. Rename the compediuns
Useful for git problems with development of foundry compediuns for system, the command git status
constantly updates the compedium files, moreover, it's more organized. All data for Foundry are
converted for JSON files in SRC folder, separeted with lang, and the DB files (JSON files converted)
are now in packs folder.