feat(item): added an chatListener, an chatCardAction Handler and gets the item information
_onChatCardAction manipulates the actions of the template's buttons by calling several other
methods. _getChatCardActor pulls the token information, if not, the actor information. getChatData
prepares information to be displayed in the chat. rollAttack/rollDamage/rollFormula perform data
rolls based on the data entered into the item. Additionally, some more item information is displayed
on the item, such as proficiency and the wielding.