feat(open source acessibility): several variables were changed. To improve code accessibility
Skill and attribute variables have been changed to English instead of Portuguese in several parts of
the system.

BREAKING CHANGE: Changing the name of the variables related to the agent form results in data loss
and thus a loss of compatibility with previous forms, it will be necessary to create another one
with the same data as before.