feat(application and roll configuration): based on the dnd5e system, were created some apps/dialogs
To implement the new applicationV2 of FoundryVTT, and the new system for rolling dice with skills,
this system is based on the dnd5e system, including an application mixin (appV2 with Handlebars) and
some dialogs with extensions. The main dialog for rolls is roll-configuration-dialog, which extends
dialogOP, and d20-configuration-dialog, and skill-tool-configuration-dialog have some polymorphic
methods.

BREAKING CHANGE: The ApplicationV2 implementation will close the supporte for V11 core version of
Foundry.